fix compiler warnings
git-svn-id: file:///Users/braun/svn/vermont/branches/vermont-dynamic-config@2032 aef3b71b-58ee-0310-9ba9-8811b9f0742f

+++ concentrator/IpfixSampler.cpp
@@ -29,7 +29,7 @@ IpfixSampler::IpfixSampler(double flowrate)
 	  statDropped(0),
 	  statTotalDropped(0)
 {
-	modulo = round(1.0/flowRate);
+	modulo = (uint64_t)round(1.0/flowRate);
 
 	msg(MSG_INFO, "IpfixSampler started with following parameters:");
 	msg(MSG_INFO, "  - flowRate=%f", flowRate);
